auth

+++ routes/database/migrations/2019_10_12_072241_create_road_types_table.php
@@ -16,7 +16,7 @@ class CreateRoadTypesTable extends Migration
         Schema::create('road_types', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('type_name');
-            $table->integer('delay_factor');
+            $table->decimal('delay_factor', 4, 2);
             $table->timestamps();
         });
     }
